requirements.txt
# ==============================================================================
# tideon.ai Framework - Requirements
# Python version: >=3.10
# ==============================================================================

# ==============================================================================
# Deep Learning & Model Support
# ==============================================================================
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0
sentencepiece>=0.1.99
protobuf>=3.20.0
bitsandbytes>=0.46.1

# vLLM for fast inference (5-10x faster, requires CUDA)
vllm>=0.14.0

# ==============================================================================
# API Model Adapters
# ==============================================================================
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# ==============================================================================
# Configuration Management
# ==============================================================================
pyyaml>=6.0.1
jsonschema>=4.19.0
json5>=0.9.14
python-dotenv>=1.0.0

# ==============================================================================
# Tool Support & HTTP
# ==============================================================================
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# ==============================================================================
# API Server
# ==============================================================================
fastapi>=0.109.0
uvicorn>=0.27.0

# ==============================================================================
# Web Search
# ==============================================================================
duckduckgo-search>=8.1.0
ddgs>=9.9.3

# ==============================================================================
# File Operations & Code Execution
# ==============================================================================
docker>=6.1.0
nbformat>=5.9.0
jupyter-client>=8.5.0

# ==============================================================================
# Utilities
# ==============================================================================
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
tqdm>=4.66.0
rich>=13.7.0
click>=8.1.0
colorama>=0.4.6
jinja2>=3.1.0

# ==============================================================================
# Async Support
# ==============================================================================
asyncio-throttle>=1.0.1
nest-asyncio>=1.5.8

# ==============================================================================
# Logging & Monitoring
# ==============================================================================
loguru>=0.7.0

# ==============================================================================
# GPU Management
# ==============================================================================
gputil>=1.4.0
nvidia-ml-py>=11.495.46

# ==============================================================================
# Protocol Support - MCP (Model Context Protocol) Official SDK
# ==============================================================================
mcp>=1.25.0

# ==============================================================================
# Performance Monitoring
# ==============================================================================
psutil>=5.9.0

# ==============================================================================
# Data Processing
# ==============================================================================
tiktoken>=0.5.0
scikit-learn>=1.3.0
datasets>=2.16.0
sentence-transformers>=2.2.0

# ==============================================================================
# Security
# ==============================================================================
cryptography>=41.0.0

# ==============================================================================
# Testing (minimal, see requirements-dev.txt for full suite)
# ==============================================================================
pytest>=7.4.0
